+++ src/routes/index.tsx
@@ -102,8 +102,18 @@ export default function Home() {
       <Motion.section
         initial={{ transform: "scaleY(0)" }}
         inView={{ transform: "scaleY(1)" }}
-        class="h-[100vh] w-full flex flex-col items-center justify-center  bg-primary text-background"
+        class="relative h-[100vh] w-full flex flex-col items-center justify-center  bg-primary text-background"
       >
+        
+        <a href="/autozap/" class="z-10 text-white justify-center flex flex-col items-center text-center font-bold rotate-12 absolute w-[10rem] h-[10rem] bg-black right-[1rem] top-[1rem] rounded-full">
+          Check out <span class="underline font-black">
+            AutoZaps!
+            </span>
+        </a>
+        <div class="animate-[shadowbounce_2s_ease-in-out_infinite] z-5 absolute w-[10rem] h-[10rem] bg-black/30 right-[0.75rem] top-[1.25rem] rounded-full">
+          </div>
+        
+        
         <Motion.img
           src="/zapple-logo.svg"
           class="mx-auto h-[270px] w-[270px]"
